+++ tests/Doubles/Laravel9/Application.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Doubles\Laravel9;
+namespace Imdhemy\Purchases\Tests\Doubles\Laravel9;
 
 use Illuminate\Foundation\Application as IlluminateApplication;