+++ tests/Integration/Artisan/ArtisanCommandsTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Integration\Artisan;
+namespace Imdhemy\Purchases\Tests\Integration\Artisan;
 
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Console\Kernel;
@@ -11,8 +11,8 @@ use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 use Imdhemy\Purchases\Console\LiapConfigPublishCommand;
 use Imdhemy\Purchases\Console\LiapUrlCommand;
-use Tests\Doubles\UrlGeneratorDouble;
-use Tests\TestCase;
+use Imdhemy\Purchases\Tests\Doubles\UrlGeneratorDouble;
+use Imdhemy\Purchases\Tests\TestCase;
 
 /**
  * The word `it` refers to LIAP package.